refactor: separate kube utils package
