+++ pkg/common/common.go
@@ -70,21 +70,6 @@ func GetNamespaceLabels(namespaceObj *v1.Namespace, logger logr.Logger) map[stri
 	return namespaceUnstructured.GetLabels()
 }
 
-// GetKindFromGVK - get kind and APIVersion from GVK
-func GetKindFromGVK(str string) (apiVersion string, kind string) {
-	if strings.Count(str, "/") == 0 {
-		return "", str
-	}
-	splitString := strings.Split(str, "/")
-	if strings.Count(str, "/") == 1 {
-		return splitString[0], splitString[1]
-	}
-	if splitString[1] == "*" {
-		return "", splitString[2]
-	}
-	return splitString[0] + "/" + splitString[1], splitString[2]
-}
-
 func VariableToJSON(key, value string) []byte {
 	var subString string
 	splitBySlash := strings.Split(key, "\"")
@@ -202,15 +187,4 @@ func removePolicyFromLabels(pName string, labels map[string]string) (bool, map[s
 	}
 
 	return false, labels
-}
-
-func GetFormatedKind(str string) (kind string) {
-	if strings.Count(str, "/") == 0 {
-		return strings.Title(str)
-	}
-	splitString := strings.Split(str, "/")
-	if strings.Count(str, "/") == 1 {
-		return splitString[0] + "/" + strings.Title(splitString[1])
-	}
-	return splitString[0] + "/" + splitString[1] + "/" + strings.Title(splitString[2])
 }
